requirements.txt
fastapi
uvicorn[standard]
python-socketio[asyncio]
python-engineio @ git+https://github.com/miguelgrinberg/python-engineio